Allow setting reconnect delay for `MockAgent` _after_ is has connected
As it was before, `reconnectDelay` was read and cached at connect time. That's makes testing more difficult, e.g. I create a `MockAgent` in test suite's `beforeEach` hook and connect it. Then, in some specific test, I want to set specific `reconnectDelay`, but that wouldn't work.

Now `reconnectDelay` will be read just before using it, so if value was set between connecting and reconnecting, it'll be used.